+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.6.0
+ - add -BSE Board stop effect
+        possibility to set effect when the board stops during the match
+ - add -TE Takeout effect
+        posibillity to set effect when you takeout the darts or takeout is wrongly triggert
+ - add -CE calibration effect
+        possibility to set effect which will be used when calibration is in progress
+
 ## 1.5.3
 - bugfix board takeout freeze after winning the match with arg -BSW set to 0
 - set -BSW to default true